Refactor events into their own lifecycle handler.
